requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Diffusion models
diffusers>=0.25.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.6.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0

# Metrics
torchmetrics>=1.0.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
